Reorganize 1. Rename MonadFindBranches to FindBranches 2. Move files to Git/Stale/ in anticipation of new functionality
