+++ src/Git/Stale/Types/Error.hs
@@ -1,0 +1,29 @@
+-- |
+-- Module      : Git.Stale.Types.Error
+-- License     : BSD3
+-- Maintainer  : tbidne@gmail.com
+-- Provides `Err` type for describing any errors encountered.
+module Git.Stale.Types.Error
+  ( Err (..),
+    ErrOr,
+  )
+where
+
+import qualified Data.Text as T
+
+-- | Wraps `T.Text` to describe an error. Git* errors describe errors
+-- encountered by the underlying `Core.FindBranches.FindBranches`
+-- monad. Others describe pure errors encountered during parsing data returned
+-- by the monad.
+data Err
+  = ParseLog T.Text
+  | ParseDate T.Text
+  | ParseName T.Text
+  | ParseMerge T.Text
+  | ReadInt T.Text
+  | GitBranches T.Text
+  | GitLog T.Text
+  deriving (Show)
+
+-- | Alias for convenience.
+type ErrOr a = Either Err a
